+++ src/Opaleye/Unpackspec.hs
@@ -1,3 +1,0 @@
-module Opaleye.Unpackspec (Unpackspec, unpackspecField) where
-
-import Opaleye.Internal.Unpackspec